pyproject.toml
[tool.poetry]
name = "rapport"
version = "0.1.0"
description = "For my mom"
authors = ["earthlyz9<earthlyz9.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.0"
django = "^4.2.3"
django-ninja = "^0.22.2"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
mysqlclient = "^2.2.0"
django-ninja-jwt = "^5.2.5"
six = "^1.16.0"
pre-commit = "^3.3.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"